Fix clippy warnings: remove unused imports, dead code, unnecessary casts
- Remove unused import warn from npu.rs
- Remove unused import Path from pipeline_snapshot.rs (only used in mnn cfg)
- Prefix unused struct fields with underscore to indicate intentional non-use
- Add allow(dead_code) for stub functions and constants
- Remove unnecessary usize cast in temporal/extractors.rs
- Simplify deref to use auto-deref
- Add Default impl for QuantizationManager
- Remove unnecessary mut on job parameter

All 617 tests pass, zero clippy warnings.

+++ cam-rust/cam-isp/src/temporal/extractors.rs
@@ -200,7 +200,7 @@ impl CoeffExtractor for GainExtractor {
         let mut gain_map = vec![1.0f32; size];
         for y in 0..h {
             for x in 0..w {
-                let ch = (x % 3) as usize;
+                let ch = x % 3;
                 gain_map[y * w + x] = match ch {
                     0 => r_gain,
                     1 => g_gain,
